+++ app/Http/Controllers/RolesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
@@ -17,12 +18,18 @@ class RolesController extends Controller
     //角色管理列表
     public function index()
     {
+        if (!Auth::user()->can('RBAC管理')){
+            return view('403');
+        }
         $roles = Role::paginate(10);
         return view('Roles.index',compact('roles'));
     }
     //添加角色
     public function create()
     {
+        if (!Auth::user()->can('RBAC管理')){
+            return view('403');
+        }
         $permissions = Permission::all();
         return view('Roles/create',compact('permissions'));
     }
@@ -47,12 +54,18 @@ class RolesController extends Controller
     //查看
     public function show(Role $role)
     {
+        if (!Auth::user()->can('RBAC管理')){
+            return view('403');
+        }
         $permissions = Permission::all();
         return view('Roles/show',compact('role','permissions'));
     }
     //修改
     public function edit(Role $role)
     {
+        if (!Auth::user()->can('RBAC管理')){
+            return view('403');
+        }
         //获取所有的权限
         $permissions = Permission::all();
         return view('Roles/edit',compact('role','permissions'));
@@ -74,6 +87,13 @@ class RolesController extends Controller
     //删除角色
     public function destroy(Role $role)
     {
+        if (!Auth::user()->can('RBAC管理')){
+            return view('403');
+        }
+        $permissions = Permission::all();
+        if ($role->hasAllPermissions($permissions)){
+            return back()->with('danger','该角色有权限不能删除!');
+        }
         $role->delete();
         return redirect()->route('roles.index')->with('success','删除成功!');
     }
