+++ app/Http/Controllers/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Shop;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class UsersController extends Controller
 {
@@ -23,8 +24,9 @@ class UsersController extends Controller
     //修改商家账户状态
     public function no(Request $request)
     {
-
-
+        if (!Auth::user()->can('商家账户修改')){
+            return view('403');
+        }
         if($request->status == 0){
             User::where('id','=',$request->id)->first()->update(['status'=>1]);
         }elseif ($request->status == 1){
@@ -35,6 +37,9 @@ class UsersController extends Controller
     //删除账户
     public function destroy(User $user)
     {
+        if (!Auth::user()->can('商家账户删除')){
+            return view('403');
+        }
         Shop::where('id','=',$user->shop_id)->delete();
         $user->delete();
         return redirect()->route('users.index')->with('success','删除成功!');
@@ -42,6 +47,9 @@ class UsersController extends Controller
     //重置账户密码
     public function set(User $user)
     {
+        if (!Auth::user()->can('商家账户修改')){
+            return view('403');
+        }
         return view('Users/password',compact('user'));
     }
